更新springcloud2023.0.0 springcloudalibaba2022.0.0-RC2 jdk17

+++ cart-service/src/main/java/com/hmall/cart/service/impl/CartServiceImpl.java
@@ -124,14 +124,14 @@ public class CartServiceImpl extends ServiceImpl<CartMapper, Cart> implements Ca
     }
 
     private void checkCartsFull(Long userId) {
-        int count = lambdaQuery().eq(Cart::getUserId, userId).count();
+        Long count = lambdaQuery().eq(Cart::getUserId, userId).count();
         if (count >= cartProperties.getMaxItem()) {
             throw new BizIllegalException(StrUtil.format("用户购物车课程不能超过{}", cartProperties.getMaxItem()));
         }
     }
 
     private boolean checkItemExists(Long itemId, Long userId) {
-        int count = lambdaQuery()
+        Long count = lambdaQuery()
                 .eq(Cart::getUserId, userId)
                 .eq(Cart::getItemId, itemId)
                 .count();
